+++ animation_nodes/nodes/gpencil/offset_gp_layer_frames.py
@@ -12,7 +12,7 @@ class OffsetGPLayerFramesNode(bpy.types.Node, AnimationNode):
     def create(self):
         self.newInput(VectorizedSocket("GPLayer", "useLayerList",
             ("Layer", "layer"), ("Layers", "layers")), dataIsModified = True)
-        self.newInput(VectorizedSocket("Float", "useFrameNumberList",
+        self.newInput(VectorizedSocket("Integer", "useFrameNumberList",
             ("Offset", "offset"), (" Offsets", "offsets")), value = 0)
         self.newOutput(VectorizedSocket("GPLayer", ["useLayerList", "useFrameNumberList"],
             ("Layer", "layer"), ("Layers", "outLayers")))
